Вынес все с мейна
